feat: remove lru-cache and move wrapFetch to shared package (#2796)
