+++ .changeset/smooth-pillows-jam.md
@@ -1,0 +1,6 @@
+---
+"qiankun": patch
+"@qiankunjs/shared": patch
+---
+
+feat: remove lru-cache and move wrapFetch to shared package
