+++ packages/shared/src/index.ts
@@ -7,3 +7,4 @@ export * from './utils';
 export * from './module-resolver';
 export * from './common';
 export * from './error';
+export * from './fetch-utils/wrapFetchWithCache';
